rename and move component files
